+++ fabric-renderer-registries-v1/src/testmod/java/net/fabricmc/fabric/test/renderer/registry/common/cooldown/CooldownItem.java
@@ -14,52 +14,30 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.test.renderregistry.common.durabilitybar;
+package net.fabricmc.fabric.test.renderer.registry.common.cooldown;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
 import net.minecraft.world.World;
 
 /**
- * An example item for something that stores units of something (i.e. energy / mana)
- * with a capacity of 1000 units. Right-clicking with the item in the air will cycle the level.
+ * An item that causes a cooldown when it is used.
  */
-public class StorageItem extends Item {
-	private static final int CAPACITY = 1000;
+public class CooldownItem extends Item {
+	// Cooldown duration in ticks when this item is being used
+	private final int duration;
 
-	public StorageItem(Settings settings) {
+	public CooldownItem(Settings settings, int duration) {
 		super(settings);
-	}
-
-	public float getFillLevel(ItemStack stack) {
-		CompoundTag tag = stack.getTag();
-
-		if (tag == null) {
-			return 0;
-		}
-
-		return tag.getInt("current") / (float) CAPACITY;
+		this.duration = duration;
 	}
 
 	@Override
 	public TypedActionResult<ItemStack> use(World world, PlayerEntity user, Hand hand) {
-		ItemStack item = user.getStackInHand(hand).copy();
-
-		CompoundTag tag = item.getOrCreateTag();
-		int current = tag.getInt("current");
-
-		if (current >= CAPACITY) {
-			current = 0; // cycle back to 0
-		} else {
-			current = Math.min(CAPACITY, current + CAPACITY / 5);
-		}
-
-		tag.putInt("current", current);
-
-		return TypedActionResult.success(item);
+		user.getItemCooldownManager().set(this, duration);
+		return TypedActionResult.success(user.getStackInHand(hand));
 	}
 }
